fix(global): Fix auth->company_id force cast conversion (int) - Replace company default for use the constant

+++ app/Http/Controllers/Account/AccountIndexController.php
@@ -14,7 +14,7 @@ class AccountIndexController extends MainController
     public function index(Request $request)
     {
         $account = new Account();
-        $data['accounts'] = $account->getAllActiveByCompany(Auth::user()->company_id);
+        $data['accounts'] = $account->getAllActiveByCompany((int) Auth::user()->company_id);
 
         return view('account.index', $data);
     }
